requirements.txt
Flask
openai-whisper
torch